Modify data chunk
